Move telemetry-related classes to a separate package
CHANGELOG_BEGIN
CHANGELOG_END
