+++ ledger/metrics/src/main/scala/com/daml/metrics/package.scala
@@ -5,12 +5,8 @@ package com.daml
 
 import com.codahale.metrics.MetricRegistry.MetricSupplier
 import com.codahale.metrics.{Gauge, MetricRegistry}
-import io.opentelemetry.api.GlobalOpenTelemetry
-import io.opentelemetry.api.trace.Tracer
 
 package object metrics {
-
-  val ParticipantTracer: Tracer = GlobalOpenTelemetry.getTracer("participant")
 
   private[metrics] def registerGauge(
       name: MetricName,
