+++ ledger/metrics/src/main/scala/com/daml/telemetry/Spans.scala
@@ -1,7 +1,4 @@
-// Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.daml.metrics
+package com.daml.telemetry
 
 import io.opentelemetry.api.trace.Span
